Remove support for the default methods and constructors
Actually, in all cases the modified methods and constructors should be used.
The idea behind it was, that I didn't want to break plugins for example,
which make use of that nms stuff.
On the other hand it is bad supporting the default methods and constructors
because this makes updating more error-prone.
In particular it's easy to miss a new method or constructor call in an update,
which should be changed to call the modified one.